requirements.txt
asgiref==3.7.2
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
coverage==7.4.2
coverage-badge==1.1.0
cryptography==42.0.2
decorator==5.1.1
defusedxml==0.8.0rc2
dj-database-url==2.1.0
Django==5.0.2
django-templated-mail==1.1.1
djangorestframework==3.14.0

djangorestframework-simplejwt==5.3.1
djoser==2.2.2
docopt==0.6.2
drf-spectacular==0.27.1
executing==2.0.1
fastjsonschema==2.19.1
genbadge==1.1.1
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
ipython==8.12.3
itypes==1.2.0
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab_pygments==0.3.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mistune==3.0.2
nbclient==0.9.0
nbconvert==7.16.1
nbformat==5.9.2
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.2
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pipreqs==0.5.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.18.0
simplejson==3.19.2
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.2
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
tinycss2==1.2.1
tornado==6.4
traitlets==5.14.1
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.0
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
django-cors-headers==4.3.1
